requirements.txt
discord.py
grpcio
python-dotenv
requests
audioop-lts; python_version>='3.13'
